Generalize delay load thunks to also work for eager fixups
This change modifies the delay load helper thunks I previously
implemented to also optionally support eager import cells - in case
of eager import cells, the generated thunk code contains just
the indirect call through the cell, the other instructions are
not applicable in this case.

Thanks

Tomas
